requirements.txt
transformers
torch
pillow
requests
safetensors
fire
numpy